updating plugins
