+++ azure/rs_azure_lb/rs_azure_networking_plugin.rb
@@ -2,7 +2,7 @@ name 'rs_azure_networking_plugin'
 type 'plugin'
 rs_ca_ver 20161221
 short_description "Azure Networking Plugin"
-long_description "Version: 1.2"
+long_description "Version: 1.3"
 package "plugins/rs_azure_networking_plugin"
 import "sys_log"
 
@@ -16,7 +16,7 @@ permission "read_creds" do
   resources "rs_cm.credentials"
 end
 
-plugin "rs_azure_networking_interfaces" do
+plugin "rs_azure_networking" do
   endpoint do
     default_host "https://management.azure.com/"
     default_scheme "https"
@@ -31,8 +31,8 @@ plugin "rs_azure_networking_interfaces" do
   end
 
   type "interface" do
-    href_templates "{{id}}","{{value[*].id}}"
-    provision "provision_resource"
+    href_templates "{{type=='Microsoft.Network/NetworkInterfaces' && id || null}}","{{value[0].type=='Microsoft.Network/NetworkInterfaces' && value[*].id || null}}"
+    provision "provision_interface"
     delete    "delete_resource"
 
     field "properties" do
@@ -111,24 +111,9 @@ plugin "rs_azure_networking_interfaces" do
 
     output "id","name","location","tags","properties"
   end
-end
-
-plugin "rs_azure_lb" do
-  endpoint do
-    default_host "https://management.azure.com/"
-    default_scheme "https"
-    query do {
-      'api-version' =>  '2016-09-01'
-    } end
-  end
-
-  parameter "subscription_id" do
-    type  "string"
-    label "subscription_id"
-  end
 
   type "load_balancer" do
-    href_templates "{{id}}"
+    href_templates "{{contains(id, 'Microsoft.Network/loadBalancers') && id || null}}"
     provision "provision_lb"
     delete    "delete_resource"
 
@@ -247,8 +232,8 @@ plugin "rs_azure_lb" do
   end
 end
 
-resource_pool "rs_azure_networking_interfaces" do
-    plugin $rs_azure_networking_interfaces
+resource_pool "rs_azure_networking" do
+    plugin $rs_azure_networking
     parameter_values do
       subscription_id $subscription_id
     end
@@ -260,24 +245,6 @@ resource_pool "rs_azure_networking_interfaces" do
         client_secret cred("AZURE_APPLICATION_KEY")
         additional_params do {
           "resource" => "https://management.azure.com/"
-        } end
-      end
-    end
-end
-
-resource_pool "rs_azure_lb" do
-    plugin $rs_azure_lb
-    parameter_values do
-      subscription_id $subscription_id
-    end
-
-    auth "azure_auth", type: "oauth2" do
-      token_url "https://login.microsoftonline.com/09b8fec1-4b8d-48dd-8afa-5c1a775ea0f2/oauth2/token"
-      grant type: "client_credentials" do
-        client_id cred("AZURE_APPLICATION_ID")
-        client_secret cred("AZURE_APPLICATION_KEY")
-        additional_params do {
-          "resource" => "https://management.azure.com/"     
         } end
       end
     end
@@ -336,7 +303,7 @@ define provision_lb(@declaration) return @resource do
   end
 end
 
-define provision_resource(@declaration) return @resource do
+define provision_interface(@declaration) return @resource do
   sub on_error: stop_debugging() do
     $object = to_object(@declaration)
     $fields = $object["fields"]
